requirements.txt
adbutils==2.7.2
androguard==3.4.0a1
frida==16.1.5
loguru==0.7.2
mitmproxy==10.4.2
pandas==2.2.3
nltk==3.9.1
python-magic-bin==0.4.14
openai>=0.27.9